Enable preview language features in project
Updated the `<LangVersion>` property in `Directory.Build.props`
from `14.0` to `preview` to allow the use of preview language
features. This enables developers to experiment with and adopt
new features that are not yet finalized in the stable version
of the language. Note that using preview features may introduce
risks if they change or are removed in future releases.